tutorials/r4hisllm/paraglidable_service/requirements.txt
fastapi 
uvicorn 
requests 
python-dotenv
aiohttp
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-requests
opentelemetry-exporter-prometheus
opentelemetry-instrumentation-aiohttp-client